refactor: rewrite code on oop concept
