+++ public/index.html
@@ -11,5 +11,5 @@
     <title>Virtual Keyboard</title>
 </head>
 <body>
-<script defer src="main.0b17aca1737f8f26afb7.js"></script></body>
+<script defer src="main.50b2888be56d947997c1.js"></script></body>
 </html>
