+++ public/main.50b2888be56d947997c1.js
@@ -1,0 +1,784 @@
+/******/ (() => { // webpackBootstrap
+/******/ 	"use strict";
+/******/ 	var __webpack_modules__ = ({
+
+/***/ "./node_modules/css-loader/dist/cjs.js!./node_modules/sass-loader/dist/cjs.js!./src/style.scss":
+/*!*****************************************************************************************************!*\
+  !*** ./node_modules/css-loader/dist/cjs.js!./node_modules/sass-loader/dist/cjs.js!./src/style.scss ***!
+  \*****************************************************************************************************/
+/***/ ((module, __webpack_exports__, __webpack_require__) => {
+
+__webpack_require__.r(__webpack_exports__);
+/* harmony export */ __webpack_require__.d(__webpack_exports__, {
+/* harmony export */   "default": () => (__WEBPACK_DEFAULT_EXPORT__)
+/* harmony export */ });
+/* harmony import */ var _node_modules_css_loader_dist_runtime_sourceMaps_js__WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! ../node_modules/css-loader/dist/runtime/sourceMaps.js */ "./node_modules/css-loader/dist/runtime/sourceMaps.js");
+/* harmony import */ var _node_modules_css_loader_dist_runtime_sourceMaps_js__WEBPACK_IMPORTED_MODULE_0___default = /*#__PURE__*/__webpack_require__.n(_node_modules_css_loader_dist_runtime_sourceMaps_js__WEBPACK_IMPORTED_MODULE_0__);
+/* harmony import */ var _node_modules_css_loader_dist_runtime_api_js__WEBPACK_IMPORTED_MODULE_1__ = __webpack_require__(/*! ../node_modules/css-loader/dist/runtime/api.js */ "./node_modules/css-loader/dist/runtime/api.js");
+/* harmony import */ var _node_modules_css_loader_dist_runtime_api_js__WEBPACK_IMPORTED_MODULE_1___default = /*#__PURE__*/__webpack_require__.n(_node_modules_css_loader_dist_runtime_api_js__WEBPACK_IMPORTED_MODULE_1__);
+// Imports
+
+
+var ___CSS_LOADER_EXPORT___ = _node_modules_css_loader_dist_runtime_api_js__WEBPACK_IMPORTED_MODULE_1___default()((_node_modules_css_loader_dist_runtime_sourceMaps_js__WEBPACK_IMPORTED_MODULE_0___default()));
+// Module
+___CSS_LOADER_EXPORT___.push([module.id, "* {\n  margin: 0;\n  padding: 0;\n  box-sizing: border-box; }\n\nbody {\n  height: 100vh;\n  width: 100vw;\n  background-color: #f5f5f5;\n  font-family: monospace;\n  font-size: 20px;\n  color: #090808; }\n\n.main-container {\n  max-width: 798px;\n  width: 100%;\n  height: auto;\n  margin: 0 auto;\n  padding: 25px 50px;\n  display: flex;\n  flex-flow: column;\n  align-items: center;\n  gap: 18px;\n  justify-content: center; }\n  .main-container textarea {\n    width: 100%;\n    height: 300px;\n    margin: 0 auto;\n    padding: 5px;\n    font: inherit;\n    resize: none;\n    outline: 0;\n    border: 3px solid #705f5f;\n    border-radius: 16px;\n    background-color: #ffffff; }\n  .main-container .keyboard {\n    width: 698px;\n    display: flex;\n    flex-wrap: wrap;\n    gap: 7px;\n    justify-content: center;\n    background-color: #ffffff; }\n    .main-container .keyboard .key {\n      display: flex;\n      justify-content: center;\n      align-items: center;\n      height: 40px;\n      width: 40px;\n      font: inherit;\n      font-family: sans-serif;\n      cursor: pointer;\n      transition: 0.3s all;\n      border: 2px solid #625353;\n      border-radius: 10%;\n      background-color: #fafaf7; }\n      .main-container .keyboard .key.active {\n        transform: scale(1);\n        border: 2px solid #625353;\n        text-align: center;\n        background-color: #cbcbb8;\n        font-size: 22px;\n        color: #112691; }\n      .main-container .keyboard .key.Backspace, .main-container .keyboard .key.ShiftLeft, .main-container .keyboard .key.ShiftRight, .main-container .keyboard .key.CapsLock, .main-container .keyboard .key.Enter {\n        width: 87px; }\n      .main-container .keyboard .key.Space {\n        width: 322px; }\n  .main-container p {\n    font-size: 22px; }\n", "",{"version":3,"sources":["webpack://./src/style.scss"],"names":[],"mappings":"AAAA;EACE,SAAS;EACT,UAAU;EACV,sBAAsB,EAAA;;AAGxB;EACE,aAAa;EACb,YAAY;EACZ,yBAAyB;EAEvB,sBAAiB;EACjB,eAAU;EAEZ,cAAc,EAAA;;AAGhB;EACE,gBAAuC;EACvC,WAAW;EACX,YAAY;EACZ,cAAc;EACd,kBAAkB;EAClB,aAAa;EACb,iBAAiB;EACjB,mBAAmB;EACnB,SAAS;EACT,uBAAuB,EAAA;EAVzB;IAaI,WAAW;IACX,aAAa;IACb,cAAc;IACd,YAAY;IACZ,aAAa;IACb,YAAY;IACZ,UAAU;IACV,yBAAyB;IACzB,mBAAmB;IACnB,yBAAyB,EAAA;EAtB7B;IA0BI,YAA2B;IAC3B,aAAa;IACb,eAAe;IACf,QAAQ;IACR,uBAAuB;IACvB,yBAAyB,EAAA;IA/B7B;MAkCM,aAAa;MACb,uBAAuB;MACvB,mBAAmB;MACnB,YAAY;MACZ,WAAW;MACX,aAAa;MACb,uBAAuB;MACvB,eAAe;MACf,oBAAoB;MACpB,yBAAyB;MACzB,kBAAkB;MAClB,yBAAyB,EAAA;MA7C/B;QAgDQ,mBAAmB;QACnB,yBAAyB;QACzB,kBAAkB;QAClB,yBAAyB;QACzB,eAAe;QACf,cAAc,EAAA;MArDtB;QA6DQ,WAAW,EAAA;MA7DnB;QAiEQ,YAAyB,EAAA;EAjEjC;IAuEI,eAAe,EAAA","sourcesContent":["* {\n  margin: 0;\n  padding: 0;\n  box-sizing: border-box;\n}\n\nbody {\n  height: 100vh;\n  width: 100vw;\n  background-color: #f5f5f5;\n  font: {\n    family: monospace;\n    size: 20px;\n  }\n  color: #090808;\n}\n\n.main-container {\n  max-width: 40px * 15 + 7px * 14 + 100px;\n  width: 100%;\n  height: auto;\n  margin: 0 auto;\n  padding: 25px 50px;\n  display: flex;\n  flex-flow: column;\n  align-items: center;\n  gap: 18px;\n  justify-content: center;\n\n  textarea {\n    width: 100%;\n    height: 300px;\n    margin: 0 auto;\n    padding: 5px;\n    font: inherit;\n    resize: none;\n    outline: 0;\n    border: 3px solid #705f5f;\n    border-radius: 16px;\n    background-color: #ffffff;\n  }\n\n  .keyboard {\n    width: 40px * 15 + 7px * 14;\n    display: flex;\n    flex-wrap: wrap;\n    gap: 7px;\n    justify-content: center;\n    background-color: #ffffff;\n\n    .key {\n      display: flex;\n      justify-content: center;\n      align-items: center;\n      height: 40px;\n      width: 40px;\n      font: inherit;\n      font-family: sans-serif;\n      cursor: pointer;\n      transition: 0.3s all;\n      border: 2px solid #625353;\n      border-radius: 10%;\n      background-color: #fafaf7;\n\n      &.active {\n        transform: scale(1);\n        border: 2px solid #625353;\n        text-align: center;\n        background-color: #cbcbb8;\n        font-size: 22px;\n        color: #112691;\n      }\n\n      &.Backspace,\n      &.ShiftLeft,\n      &.ShiftRight,\n      &.CapsLock,\n      &.Enter {\n        width: 87px;\n      }\n\n      &.Space {\n        width: 40px * 7 + 7px * 6;\n      }\n    }\n  }\n\n  p {\n    font-size: 22px;\n  }\n}\n"],"sourceRoot":""}]);
+// Exports
+/* harmony default export */ const __WEBPACK_DEFAULT_EXPORT__ = (___CSS_LOADER_EXPORT___);
+
+
+/***/ }),
+
+/***/ "./node_modules/css-loader/dist/runtime/api.js":
+/*!*****************************************************!*\
+  !*** ./node_modules/css-loader/dist/runtime/api.js ***!
+  \*****************************************************/
+/***/ ((module) => {
+
+
+
+/*
+  MIT License http://www.opensource.org/licenses/mit-license.php
+  Author Tobias Koppers @sokra
+*/
+module.exports = function (cssWithMappingToString) {
+  var list = []; // return the list of modules as css string
+
+  list.toString = function toString() {
+    return this.map(function (item) {
+      var content = "";
+      var needLayer = typeof item[5] !== "undefined";
+
+      if (item[4]) {
+        content += "@supports (".concat(item[4], ") {");
+      }
+
+      if (item[2]) {
+        content += "@media ".concat(item[2], " {");
+      }
+
+      if (needLayer) {
+        content += "@layer".concat(item[5].length > 0 ? " ".concat(item[5]) : "", " {");
+      }
+
+      content += cssWithMappingToString(item);
+
+      if (needLayer) {
+        content += "}";
+      }
+
+      if (item[2]) {
+        content += "}";
+      }
+
+      if (item[4]) {
+        content += "}";
+      }
+
+      return content;
+    }).join("");
+  }; // import a list of modules into the list
+
+
+  list.i = function i(modules, media, dedupe, supports, layer) {
+    if (typeof modules === "string") {
+      modules = [[null, modules, undefined]];
+    }
+
+    var alreadyImportedModules = {};
+
+    if (dedupe) {
+      for (var k = 0; k < this.length; k++) {
+        var id = this[k][0];
+
+        if (id != null) {
+          alreadyImportedModules[id] = true;
+        }
+      }
+    }
+
+    for (var _k = 0; _k < modules.length; _k++) {
+      var item = [].concat(modules[_k]);
+
+      if (dedupe && alreadyImportedModules[item[0]]) {
+        continue;
+      }
+
+      if (typeof layer !== "undefined") {
+        if (typeof item[5] === "undefined") {
+          item[5] = layer;
+        } else {
+          item[1] = "@layer".concat(item[5].length > 0 ? " ".concat(item[5]) : "", " {").concat(item[1], "}");
+          item[5] = layer;
+        }
+      }
+
+      if (media) {
+        if (!item[2]) {
+          item[2] = media;
+        } else {
+          item[1] = "@media ".concat(item[2], " {").concat(item[1], "}");
+          item[2] = media;
+        }
+      }
+
+      if (supports) {
+        if (!item[4]) {
+          item[4] = "".concat(supports);
+        } else {
+          item[1] = "@supports (".concat(item[4], ") {").concat(item[1], "}");
+          item[4] = supports;
+        }
+      }
+
+      list.push(item);
+    }
+  };
+
+  return list;
+};
+
+/***/ }),
+
+/***/ "./node_modules/css-loader/dist/runtime/sourceMaps.js":
+/*!************************************************************!*\
+  !*** ./node_modules/css-loader/dist/runtime/sourceMaps.js ***!
+  \************************************************************/
+/***/ ((module) => {
+
+
+
+module.exports = function (item) {
+  var content = item[1];
+  var cssMapping = item[3];
+
+  if (!cssMapping) {
+    return content;
+  }
+
+  if (typeof btoa === "function") {
+    var base64 = btoa(unescape(encodeURIComponent(JSON.stringify(cssMapping))));
+    var data = "sourceMappingURL=data:application/json;charset=utf-8;base64,".concat(base64);
+    var sourceMapping = "/*# ".concat(data, " */");
+    var sourceURLs = cssMapping.sources.map(function (source) {
+      return "/*# sourceURL=".concat(cssMapping.sourceRoot || "").concat(source, " */");
+    });
+    return [content].concat(sourceURLs).concat([sourceMapping]).join("\n");
+  }
+
+  return [content].join("\n");
+};
+
+/***/ }),
+
+/***/ "./src/style.scss":
+/*!************************!*\
+  !*** ./src/style.scss ***!
+  \************************/
+/***/ ((__unused_webpack_module, __webpack_exports__, __webpack_require__) => {
+
+__webpack_require__.r(__webpack_exports__);
+/* harmony export */ __webpack_require__.d(__webpack_exports__, {
+/* harmony export */   "default": () => (__WEBPACK_DEFAULT_EXPORT__)
+/* harmony export */ });
+/* harmony import */ var _node_modules_style_loader_dist_runtime_injectStylesIntoStyleTag_js__WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! !../node_modules/style-loader/dist/runtime/injectStylesIntoStyleTag.js */ "./node_modules/style-loader/dist/runtime/injectStylesIntoStyleTag.js");
+/* harmony import */ var _node_modules_style_loader_dist_runtime_injectStylesIntoStyleTag_js__WEBPACK_IMPORTED_MODULE_0___default = /*#__PURE__*/__webpack_require__.n(_node_modules_style_loader_dist_runtime_injectStylesIntoStyleTag_js__WEBPACK_IMPORTED_MODULE_0__);
+/* harmony import */ var _node_modules_style_loader_dist_runtime_styleDomAPI_js__WEBPACK_IMPORTED_MODULE_1__ = __webpack_require__(/*! !../node_modules/style-loader/dist/runtime/styleDomAPI.js */ "./node_modules/style-loader/dist/runtime/styleDomAPI.js");
+/* harmony import */ var _node_modules_style_loader_dist_runtime_styleDomAPI_js__WEBPACK_IMPORTED_MODULE_1___default = /*#__PURE__*/__webpack_require__.n(_node_modules_style_loader_dist_runtime_styleDomAPI_js__WEBPACK_IMPORTED_MODULE_1__);
+/* harmony import */ var _node_modules_style_loader_dist_runtime_insertBySelector_js__WEBPACK_IMPORTED_MODULE_2__ = __webpack_require__(/*! !../node_modules/style-loader/dist/runtime/insertBySelector.js */ "./node_modules/style-loader/dist/runtime/insertBySelector.js");
+/* harmony import */ var _node_modules_style_loader_dist_runtime_insertBySelector_js__WEBPACK_IMPORTED_MODULE_2___default = /*#__PURE__*/__webpack_require__.n(_node_modules_style_loader_dist_runtime_insertBySelector_js__WEBPACK_IMPORTED_MODULE_2__);
+/* harmony import */ var _node_modules_style_loader_dist_runtime_setAttributesWithoutAttributes_js__WEBPACK_IMPORTED_MODULE_3__ = __webpack_require__(/*! !../node_modules/style-loader/dist/runtime/setAttributesWithoutAttributes.js */ "./node_modules/style-loader/dist/runtime/setAttributesWithoutAttributes.js");
+/* harmony import */ var _node_modules_style_loader_dist_runtime_setAttributesWithoutAttributes_js__WEBPACK_IMPORTED_MODULE_3___default = /*#__PURE__*/__webpack_require__.n(_node_modules_style_loader_dist_runtime_setAttributesWithoutAttributes_js__WEBPACK_IMPORTED_MODULE_3__);
+/* harmony import */ var _node_modules_style_loader_dist_runtime_insertStyleElement_js__WEBPACK_IMPORTED_MODULE_4__ = __webpack_require__(/*! !../node_modules/style-loader/dist/runtime/insertStyleElement.js */ "./node_modules/style-loader/dist/runtime/insertStyleElement.js");
+/* harmony import */ var _node_modules_style_loader_dist_runtime_insertStyleElement_js__WEBPACK_IMPORTED_MODULE_4___default = /*#__PURE__*/__webpack_require__.n(_node_modules_style_loader_dist_runtime_insertStyleElement_js__WEBPACK_IMPORTED_MODULE_4__);
+/* harmony import */ var _node_modules_style_loader_dist_runtime_styleTagTransform_js__WEBPACK_IMPORTED_MODULE_5__ = __webpack_require__(/*! !../node_modules/style-loader/dist/runtime/styleTagTransform.js */ "./node_modules/style-loader/dist/runtime/styleTagTransform.js");
+/* harmony import */ var _node_modules_style_loader_dist_runtime_styleTagTransform_js__WEBPACK_IMPORTED_MODULE_5___default = /*#__PURE__*/__webpack_require__.n(_node_modules_style_loader_dist_runtime_styleTagTransform_js__WEBPACK_IMPORTED_MODULE_5__);
+/* harmony import */ var _node_modules_css_loader_dist_cjs_js_node_modules_sass_loader_dist_cjs_js_style_scss__WEBPACK_IMPORTED_MODULE_6__ = __webpack_require__(/*! !!../node_modules/css-loader/dist/cjs.js!../node_modules/sass-loader/dist/cjs.js!./style.scss */ "./node_modules/css-loader/dist/cjs.js!./node_modules/sass-loader/dist/cjs.js!./src/style.scss");
+
+      
+      
+      
+      
+      
+      
+      
+      
+      
+
+var options = {};
+
+options.styleTagTransform = (_node_modules_style_loader_dist_runtime_styleTagTransform_js__WEBPACK_IMPORTED_MODULE_5___default());
+options.setAttributes = (_node_modules_style_loader_dist_runtime_setAttributesWithoutAttributes_js__WEBPACK_IMPORTED_MODULE_3___default());
+
+      options.insert = _node_modules_style_loader_dist_runtime_insertBySelector_js__WEBPACK_IMPORTED_MODULE_2___default().bind(null, "head");
+    
+options.domAPI = (_node_modules_style_loader_dist_runtime_styleDomAPI_js__WEBPACK_IMPORTED_MODULE_1___default());
+options.insertStyleElement = (_node_modules_style_loader_dist_runtime_insertStyleElement_js__WEBPACK_IMPORTED_MODULE_4___default());
+
+var update = _node_modules_style_loader_dist_runtime_injectStylesIntoStyleTag_js__WEBPACK_IMPORTED_MODULE_0___default()(_node_modules_css_loader_dist_cjs_js_node_modules_sass_loader_dist_cjs_js_style_scss__WEBPACK_IMPORTED_MODULE_6__["default"], options);
+
+
+
+
+       /* harmony default export */ const __WEBPACK_DEFAULT_EXPORT__ = (_node_modules_css_loader_dist_cjs_js_node_modules_sass_loader_dist_cjs_js_style_scss__WEBPACK_IMPORTED_MODULE_6__["default"] && _node_modules_css_loader_dist_cjs_js_node_modules_sass_loader_dist_cjs_js_style_scss__WEBPACK_IMPORTED_MODULE_6__["default"].locals ? _node_modules_css_loader_dist_cjs_js_node_modules_sass_loader_dist_cjs_js_style_scss__WEBPACK_IMPORTED_MODULE_6__["default"].locals : undefined);
+
+
+/***/ }),
+
+/***/ "./node_modules/style-loader/dist/runtime/injectStylesIntoStyleTag.js":
+/*!****************************************************************************!*\
+  !*** ./node_modules/style-loader/dist/runtime/injectStylesIntoStyleTag.js ***!
+  \****************************************************************************/
+/***/ ((module) => {
+
+
+
+var stylesInDOM = [];
+
+function getIndexByIdentifier(identifier) {
+  var result = -1;
+
+  for (var i = 0; i < stylesInDOM.length; i++) {
+    if (stylesInDOM[i].identifier === identifier) {
+      result = i;
+      break;
+    }
+  }
+
+  return result;
+}
+
+function modulesToDom(list, options) {
+  var idCountMap = {};
+  var identifiers = [];
+
+  for (var i = 0; i < list.length; i++) {
+    var item = list[i];
+    var id = options.base ? item[0] + options.base : item[0];
+    var count = idCountMap[id] || 0;
+    var identifier = "".concat(id, " ").concat(count);
+    idCountMap[id] = count + 1;
+    var indexByIdentifier = getIndexByIdentifier(identifier);
+    var obj = {
+      css: item[1],
+      media: item[2],
+      sourceMap: item[3],
+      supports: item[4],
+      layer: item[5]
+    };
+
+    if (indexByIdentifier !== -1) {
+      stylesInDOM[indexByIdentifier].references++;
+      stylesInDOM[indexByIdentifier].updater(obj);
+    } else {
+      var updater = addElementStyle(obj, options);
+      options.byIndex = i;
+      stylesInDOM.splice(i, 0, {
+        identifier: identifier,
+        updater: updater,
+        references: 1
+      });
+    }
+
+    identifiers.push(identifier);
+  }
+
+  return identifiers;
+}
+
+function addElementStyle(obj, options) {
+  var api = options.domAPI(options);
+  api.update(obj);
+
+  var updater = function updater(newObj) {
+    if (newObj) {
+      if (newObj.css === obj.css && newObj.media === obj.media && newObj.sourceMap === obj.sourceMap && newObj.supports === obj.supports && newObj.layer === obj.layer) {
+        return;
+      }
+
+      api.update(obj = newObj);
+    } else {
+      api.remove();
+    }
+  };
+
+  return updater;
+}
+
+module.exports = function (list, options) {
+  options = options || {};
+  list = list || [];
+  var lastIdentifiers = modulesToDom(list, options);
+  return function update(newList) {
+    newList = newList || [];
+
+    for (var i = 0; i < lastIdentifiers.length; i++) {
+      var identifier = lastIdentifiers[i];
+      var index = getIndexByIdentifier(identifier);
+      stylesInDOM[index].references--;
+    }
+
+    var newLastIdentifiers = modulesToDom(newList, options);
+
+    for (var _i = 0; _i < lastIdentifiers.length; _i++) {
+      var _identifier = lastIdentifiers[_i];
+
+      var _index = getIndexByIdentifier(_identifier);
+
+      if (stylesInDOM[_index].references === 0) {
+        stylesInDOM[_index].updater();
+
+        stylesInDOM.splice(_index, 1);
+      }
+    }
+
+    lastIdentifiers = newLastIdentifiers;
+  };
+};
+
+/***/ }),
+
+/***/ "./node_modules/style-loader/dist/runtime/insertBySelector.js":
+/*!********************************************************************!*\
+  !*** ./node_modules/style-loader/dist/runtime/insertBySelector.js ***!
+  \********************************************************************/
+/***/ ((module) => {
+
+
+
+var memo = {};
+/* istanbul ignore next  */
+
+function getTarget(target) {
+  if (typeof memo[target] === "undefined") {
+    var styleTarget = document.querySelector(target); // Special case to return head of iframe instead of iframe itself
+
+    if (window.HTMLIFrameElement && styleTarget instanceof window.HTMLIFrameElement) {
+      try {
+        // This will throw an exception if access to iframe is blocked
+        // due to cross-origin restrictions
+        styleTarget = styleTarget.contentDocument.head;
+      } catch (e) {
+        // istanbul ignore next
+        styleTarget = null;
+      }
+    }
+
+    memo[target] = styleTarget;
+  }
+
+  return memo[target];
+}
+/* istanbul ignore next  */
+
+
+function insertBySelector(insert, style) {
+  var target = getTarget(insert);
+
+  if (!target) {
+    throw new Error("Couldn't find a style target. This probably means that the value for the 'insert' parameter is invalid.");
+  }
+
+  target.appendChild(style);
+}
+
+module.exports = insertBySelector;
+
+/***/ }),
+
+/***/ "./node_modules/style-loader/dist/runtime/insertStyleElement.js":
+/*!**********************************************************************!*\
+  !*** ./node_modules/style-loader/dist/runtime/insertStyleElement.js ***!
+  \**********************************************************************/
+/***/ ((module) => {
+
+
+
+/* istanbul ignore next  */
+function insertStyleElement(options) {
+  var element = document.createElement("style");
+  options.setAttributes(element, options.attributes);
+  options.insert(element, options.options);
+  return element;
+}
+
+module.exports = insertStyleElement;
+
+/***/ }),
+
+/***/ "./node_modules/style-loader/dist/runtime/setAttributesWithoutAttributes.js":
+/*!**********************************************************************************!*\
+  !*** ./node_modules/style-loader/dist/runtime/setAttributesWithoutAttributes.js ***!
+  \**********************************************************************************/
+/***/ ((module, __unused_webpack_exports, __webpack_require__) => {
+
+
+
+/* istanbul ignore next  */
+function setAttributesWithoutAttributes(styleElement) {
+  var nonce =  true ? __webpack_require__.nc : 0;
+
+  if (nonce) {
+    styleElement.setAttribute("nonce", nonce);
+  }
+}
+
+module.exports = setAttributesWithoutAttributes;
+
+/***/ }),
+
+/***/ "./node_modules/style-loader/dist/runtime/styleDomAPI.js":
+/*!***************************************************************!*\
+  !*** ./node_modules/style-loader/dist/runtime/styleDomAPI.js ***!
+  \***************************************************************/
+/***/ ((module) => {
+
+
+
+/* istanbul ignore next  */
+function apply(styleElement, options, obj) {
+  var css = "";
+
+  if (obj.supports) {
+    css += "@supports (".concat(obj.supports, ") {");
+  }
+
+  if (obj.media) {
+    css += "@media ".concat(obj.media, " {");
+  }
+
+  var needLayer = typeof obj.layer !== "undefined";
+
+  if (needLayer) {
+    css += "@layer".concat(obj.layer.length > 0 ? " ".concat(obj.layer) : "", " {");
+  }
+
+  css += obj.css;
+
+  if (needLayer) {
+    css += "}";
+  }
+
+  if (obj.media) {
+    css += "}";
+  }
+
+  if (obj.supports) {
+    css += "}";
+  }
+
+  var sourceMap = obj.sourceMap;
+
+  if (sourceMap && typeof btoa !== "undefined") {
+    css += "\n/*# sourceMappingURL=data:application/json;base64,".concat(btoa(unescape(encodeURIComponent(JSON.stringify(sourceMap)))), " */");
+  } // For old IE
+
+  /* istanbul ignore if  */
+
+
+  options.styleTagTransform(css, styleElement, options.options);
+}
+
+function removeStyleElement(styleElement) {
+  // istanbul ignore if
+  if (styleElement.parentNode === null) {
+    return false;
+  }
+
+  styleElement.parentNode.removeChild(styleElement);
+}
+/* istanbul ignore next  */
+
+
+function domAPI(options) {
+  var styleElement = options.insertStyleElement(options);
+  return {
+    update: function update(obj) {
+      apply(styleElement, options, obj);
+    },
+    remove: function remove() {
+      removeStyleElement(styleElement);
+    }
+  };
+}
+
+module.exports = domAPI;
+
+/***/ }),
+
+/***/ "./node_modules/style-loader/dist/runtime/styleTagTransform.js":
+/*!*********************************************************************!*\
+  !*** ./node_modules/style-loader/dist/runtime/styleTagTransform.js ***!
+  \*********************************************************************/
+/***/ ((module) => {
+
+
+
+/* istanbul ignore next  */
+function styleTagTransform(css, styleElement) {
+  if (styleElement.styleSheet) {
+    styleElement.styleSheet.cssText = css;
+  } else {
+    while (styleElement.firstChild) {
+      styleElement.removeChild(styleElement.firstChild);
+    }
+
+    styleElement.appendChild(document.createTextNode(css));
+  }
+}
+
+module.exports = styleTagTransform;
+
+/***/ })
+
+/******/ 	});
+/************************************************************************/
+/******/ 	// The module cache
+/******/ 	var __webpack_module_cache__ = {};
+/******/ 	
+/******/ 	// The require function
+/******/ 	function __webpack_require__(moduleId) {
+/******/ 		// Check if module is in cache
+/******/ 		var cachedModule = __webpack_module_cache__[moduleId];
+/******/ 		if (cachedModule !== undefined) {
+/******/ 			return cachedModule.exports;
+/******/ 		}
+/******/ 		// Create a new module (and put it into the cache)
+/******/ 		var module = __webpack_module_cache__[moduleId] = {
+/******/ 			id: moduleId,
+/******/ 			// no module.loaded needed
+/******/ 			exports: {}
+/******/ 		};
+/******/ 	
+/******/ 		// Execute the module function
+/******/ 		__webpack_modules__[moduleId](module, module.exports, __webpack_require__);
+/******/ 	
+/******/ 		// Return the exports of the module
+/******/ 		return module.exports;
+/******/ 	}
+/******/ 	
+/************************************************************************/
+/******/ 	/* webpack/runtime/compat get default export */
+/******/ 	(() => {
+/******/ 		// getDefaultExport function for compatibility with non-harmony modules
+/******/ 		__webpack_require__.n = (module) => {
+/******/ 			var getter = module && module.__esModule ?
+/******/ 				() => (module['default']) :
+/******/ 				() => (module);
+/******/ 			__webpack_require__.d(getter, { a: getter });
+/******/ 			return getter;
+/******/ 		};
+/******/ 	})();
+/******/ 	
+/******/ 	/* webpack/runtime/define property getters */
+/******/ 	(() => {
+/******/ 		// define getter functions for harmony exports
+/******/ 		__webpack_require__.d = (exports, definition) => {
+/******/ 			for(var key in definition) {
+/******/ 				if(__webpack_require__.o(definition, key) && !__webpack_require__.o(exports, key)) {
+/******/ 					Object.defineProperty(exports, key, { enumerable: true, get: definition[key] });
+/******/ 				}
+/******/ 			}
+/******/ 		};
+/******/ 	})();
+/******/ 	
+/******/ 	/* webpack/runtime/hasOwnProperty shorthand */
+/******/ 	(() => {
+/******/ 		__webpack_require__.o = (obj, prop) => (Object.prototype.hasOwnProperty.call(obj, prop))
+/******/ 	})();
+/******/ 	
+/******/ 	/* webpack/runtime/make namespace object */
+/******/ 	(() => {
+/******/ 		// define __esModule on exports
+/******/ 		__webpack_require__.r = (exports) => {
+/******/ 			if(typeof Symbol !== 'undefined' && Symbol.toStringTag) {
+/******/ 				Object.defineProperty(exports, Symbol.toStringTag, { value: 'Module' });
+/******/ 			}
+/******/ 			Object.defineProperty(exports, '__esModule', { value: true });
+/******/ 		};
+/******/ 	})();
+/******/ 	
+/************************************************************************/
+var __webpack_exports__ = {};
+// This entry need to be wrapped in an IIFE because it need to be isolated against other modules in the chunk.
+(() => {
+/*!**********************!*\
+  !*** ./src/index.js ***!
+  \**********************/
+__webpack_require__.r(__webpack_exports__);
+/* harmony import */ var _style_scss__WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! ./style.scss */ "./src/style.scss");
+
+
+class Keyboard {
+    constructor(keys) {
+        this.keysElement = keys;
+        this.keyboard = null;
+        this.textarea = null;
+        this.isCapsOn = false;
+        this.isDown = false;
+        this.layout = localStorage.getItem('lang') === 'ru' ? 'ru' : 'en';
+        this.keys = [];
+    }
+
+    renderHtmlElements() {
+        const main = document.createElement('main');
+        const textarea = document.createElement('textarea');
+        const keyboard = document.createElement('div');
+        const guide = document.createElement('p');
+        const heading = document.createElement('h1');
+
+        main.classList.add('main');
+        main.classList.add('main-container');
+
+        heading.textContent = 'RSS Virtual keyboard ';
+        guide.innerText = 'To change the language, press Ctrl + Alt\nThe keyboard was written on Linux Ubuntu';
+
+        main.appendChild(heading);
+        main.appendChild(textarea);
+        main.appendChild(keyboard);
+        main.appendChild(guide);
+        document.body.appendChild(main);
+
+        this.keyboard = keyboard;
+        this.textarea = textarea;
+    }
+
+    renderKeyboard() {
+        this.keyboard.classList.add('keyboard');
+        this.keysElement.forEach((key) => {
+            const keyButton = document.createElement('button');
+            keyButton.setAttribute('id', key.code);
+            keyButton.setAttribute('type', 'button');
+            keyButton.setAttribute('character-en', key.character.en);
+            keyButton.setAttribute('character-ru', key.character.ru);
+            keyButton.setAttribute('contents-en', key.contents.en);
+            keyButton.setAttribute('contents-ru', key.contents.ru);
+            keyButton.setAttribute('shift-en', key.shift.en);
+            keyButton.setAttribute('shift-ru', key.shift.ru);
+            keyButton.classList.add('key', key.code);
+            keyButton.textContent = this.layout === 'en' ? key.contents.en : key.contents.ru;
+            this.keyboard.appendChild(keyButton);
+            this.keys.push(keyButton);
+        });
+    }
+
+    addListeners() {
+        document.addEventListener('keydown', (keyEvent) => {
+            const key = document.getElementById(keyEvent.code);
+            if (key) {
+                key.classList.add('active');
+                keyEvent.preventDefault();
+                const characterEn = key.getAttribute('character-en');
+                if ((keyEvent.code === 'ShiftLeft' || keyEvent.code === 'ShiftRight') && !keyEvent.repeat)
+                    this.shiftText(true);
+                else if (keyEvent.code === 'CapsLock' && !keyEvent.repeat) {
+                    this.shiftText(false);
+                    if (this.isCapsOn) key.classList.remove('active');
+                    this.isCapsOn = !this.isCapsOn;
+                } else if (keyEvent.ctrlKey && keyEvent.altKey) {
+                    this.layout = this.layout === 'ru' ? 'en' : 'ru';
+                    localStorage.setItem('lang', this.layout);
+                    this.keys.forEach((keys) => {
+                        if (this.layout === 'en') {
+                            const getLang = keys;
+                            getLang.innerText = keys.getAttribute('contents-en');
+                        } else if (this.layout === 'ru') {
+                            const getLang = keys;
+                            getLang.innerText = keys.getAttribute('contents-ru');
+                        }
+                    });
+                } else if (keyEvent.code === 'Backspace') this.handleText('', 'Backspace');
+                else if (keyEvent.code === 'Delete') this.handleText('', 'Delete');
+                else if (keyEvent.code === 'Enter') this.handleText('\n');
+                else if (keyEvent.code === 'Tab') this.handleText('\t');
+                else if (characterEn !== 'functional') this.handleText(key.textContent);
+            }
+        });
+
+        document.addEventListener('keyup', (keyEvent) => {
+            const key = document.getElementById(keyEvent.code);
+            if (key) {
+                if (keyEvent.code !== 'CapsLock') key.classList.remove('active');
+                if (keyEvent.code === 'ShiftLeft' || keyEvent.code === 'ShiftRight') this.shiftText(true);
+            }
+        });
+
+        this.keyboard.addEventListener('mousedown', (event) => {
+            const eventKeyDown = new KeyboardEvent('keydown', {
+                code: event.target.id,
+            });
+            document.dispatchEvent(eventKeyDown);
+            this.isDown = true;
+        });
+
+        this.keyboard.addEventListener('mouseup', (event) => {
+            const eventKeyUp = new KeyboardEvent('keyup', { code: event.target.id });
+            document.dispatchEvent(eventKeyUp);
+            this.isDown = false;
+        });
+
+        this.keyboard.addEventListener('mouseout', (event) => {
+            if (this.isDown) {
+                const eventKeyUp = new KeyboardEvent('keyup', {
+                    code: event.target.id,
+                });
+                document.dispatchEvent(eventKeyUp);
+            }
+        });
+
+        this.textarea.onblur = () => this.textarea.focus();
+        this.textarea.focus();
+    }
+
+    handleText(text, param) {
+        let cursorPos = this.textarea.selectionStart;
+        let cursorPosEnd = this.textarea.selectionEnd;
+        if (this.textarea.selectionStart === this.textarea.selectionEnd) {
+            if (param === 'Delete') cursorPosEnd += 1;
+            else if (param === 'Backspace') cursorPos = Math.max(0, cursorPos - 1);
+        }
+
+        if (param === 'Delete' || param === 'Backspace') {
+            this.textarea.setRangeText('', cursorPos, cursorPosEnd);
+        } else this.textarea.setRangeText(text);
+
+        this.textarea.selectionStart = cursorPos + text.length;
+        this.textarea.selectionEnd = this.textarea.selectionStart;
+    }
+
+    shiftText(noCapsKey) {
+        this.keys.forEach((keys) => {
+            const key = keys;
+            if (noCapsKey || key.getAttribute('character-en') === 'letter') {
+                const temp = key.getAttribute('contents-en');
+                key.setAttribute('contents-en', key.getAttribute('shift-en'));
+                key.setAttribute('shift-en', temp);
+            }
+            if (noCapsKey || key.getAttribute('character-ru') === 'letter') {
+                const temp = key.getAttribute('contents-ru');
+                key.setAttribute('contents-ru', key.getAttribute('shift-ru'));
+                key.setAttribute('shift-ru', temp);
+            }
+            key.innerText = this.layout === 'en' ? key.getAttribute('contents-en') : key.getAttribute('contents-ru');
+        });
+    }
+}
+
+const getKeysConfig = async () => {
+    const url = './assets/data/keys.json';
+    const res = await fetch(url);
+    return res.json();
+};
+
+window.addEventListener('DOMContentLoaded', async () => {
+    const keysConfig = await getKeysConfig();
+    const keyboard = new Keyboard(keysConfig);
+    keyboard.renderHtmlElements();
+    keyboard.renderKeyboard();
+    keyboard.addListeners();
+});
+
+})();
+
+/******/ })()
+;
+//# sourceMappingURL=main.50b2888be56d947997c1.js.map
