--- src/assets/js/keyMap.js
@@ -1,517 +1,0 @@
-const keyBtn = [
-    {
-        keyCode: 'Backquote',
-        ru: 'ё',
-        RU: 'Ё',
-        en: '`',
-        EN: '~',
-        perceivesCapsLk: true,
-    },
-    {
-        keyCode: 'Digit1',
-        ru: '1',
-        RU: '!',
-        en: '1',
-        EN: '!',
-        perceivesCapsLk: false,
-    },
-    {
-        keyCode: 'Digit2',
-        ru: '2',
-        RU: '"',
-        en: '2',
-        EN: '@',
-        perceivesCapsLk: false,
-    },
-    {
-        keyCode: 'Digit3',
-        ru: '3',
-        RU: '№',
-        en: '3',
-        EN: '#',
-        perceivesCapsLk: false,
-    },
-    {
-        keyCode: 'Digit4',
-        ru: '4',
-        RU: ';',
-        en: '4',
-        EN: '$',
-        perceivesCapsLk: false,
-    },
-    {
-        keyCode: 'Digit5',
-        ru: '5',
-        RU: '%',
-        en: '5',
-        EN: '%',
-        perceivesCapsLk: false,
-    },
-    {
-        keyCode: 'Digit6',
-        ru: '6',
-        RU: ':',
-        en: '6',
-        EN: '^',
-        perceivesCapsLk: false,
-    },
-    {
-        keyCode: 'Digit7',
-        ru: '7',
-        RU: '?',
-        en: '7',
-        EN: '&',
-        perceivesCapsLk: false,
-    },
-    {
-        keyCode: 'Digit8',
-        ru: '8',
-        RU: '*',
-        en: '8',
-        EN: '*',
-        perceivesCapsLk: false,
-    },
-    {
-        keyCode: 'Digit9',
-        ru: '9',
-        RU: '(',
-        en: '9',
-        EN: '(',
-        perceivesCapsLk: false,
-    },
-    {
-        keyCode: 'Digit0',
-        ru: '0',
-        RU: ')',
-        en: '0',
-        EN: ')',
-        perceivesCapsLk: false,
-    },
-    {
-        keyCode: 'Minus',
-        ru: '-',
-        RU: '_',
-        en: '-',
-        EN: '_',
-        perceivesCapsLk: false,
-    },
-    {
-        keyCode: 'Equal',
-        ru: '=',
-        RU: '+',
-        en: '=',
-        EN: '+',
-        perceivesCapsLk: false,
-    },
-    {
-        keyCode: 'Backspace',
-        ru: '⬅',
-        RU: '⬅',
-        en: '⬅',
-        EN: '⬅',
-        perceivesCapsLk: false,
-    },
-    {
-        keyCode: 'Tab',
-        ru: 'Tab',
-        RU: 'Tab',
-        en: 'Tab',
-        EN: 'Tab',
-        perceivesCapsLk: false,
-    },
-    {
-        keyCode: 'KeyQ',
-        ru: 'й',
-        RU: 'Й',
-        en: 'q',
-        EN: 'Q',
-        perceivesCapsLk: true,
-    },
-    {
-        keyCode: 'KeyW',
-        ru: 'ц',
-        RU: 'Ц',
-        en: 'w',
-        EN: 'W',
-        perceivesCapsLk: true,
-    },
-    {
-        keyCode: 'KeyE',
-        ru: 'у',
-        RU: 'У',
-        en: 'e',
-        EN: 'E',
-        perceivesCapsLk: true,
-    },
-    {
-        keyCode: 'KeyR',
-        ru: 'к',
-        RU: 'К',
-        en: 'r',
-        EN: 'R',
-        perceivesCapsLk: true,
-    },
-    {
-        keyCode: 'KeyT',
-        ru: 'е',
-        RU: 'Е',
-        en: 't',
-        EN: 'T',
-        perceivesCapsLk: true,
-    },
-    {
-        keyCode: 'KeyY',
-        ru: 'н',
-        RU: 'Н',
-        en: 'y',
-        EN: 'Y',
-        perceivesCapsLk: true,
-    },
-    {
-        keyCode: 'KeyU',
-        ru: 'г',
-        RU: 'Г',
-        en: 'u',
-        EN: 'U',
-        perceivesCapsLk: true,
-    },
-    {
-        keyCode: 'KeyI',
-        ru: 'ш',
-        RU: 'Ш',
-        en: 'i',
-        EN: 'I',
-        perceivesCapsLk: true,
-    },
-    {
-        keyCode: 'KeyO',
-        ru: 'щ',
-        RU: 'Щ',
-        en: 'o',
-        EN: 'O',
-        perceivesCapsLk: true,
-    },
-    {
-        keyCode: 'KeyP',
-        ru: 'з',
-        RU: 'З',
-        en: 'p',
-        EN: 'P',
-        perceivesCapsLk: true,
-    },
-    {
-        keyCode: 'BracketLeft',
-        ru: 'х',
-        RU: 'Х',
-        en: '[',
-        en2: false,
-        EN: '{',
-        perceivesCapsLk: true,
-    },
-    {
-        keyCode: 'BracketRight',
-        ru: 'ъ',
-        RU: 'Ъ',
-        en: ']',
-        EN: '}',
-        perceivesCapsLk: true,
-    },
-    {
-        keyCode: 'Backslash',
-        ru: '\\',
-        RU: '/',
-        en: '\\',
-        EN: '|',
-        perceivesCapsLk: false,
-    },
-    {
-        keyCode: 'Delete',
-        ru: 'Del',
-        RU: 'Del',
-        en: 'Del',
-        EN: 'Del',
-        perceivesCapsLk: false,
-    },
-    {
-        keyCode: 'CapsLock',
-        ru: 'CapsLk',
-        RU: 'CapsLk',
-        en: 'CapsLk',
-        EN: 'CapsLk',
-        perceivesCapsLk: false,
-    },
-    {
-        keyCode: 'KeyA',
-        ru: 'ф',
-        RU: 'Ф',
-        en: 'a',
-        EN: 'A',
-        perceivesCapsLk: true,
-    },
-    {
-        keyCode: 'KeyS',
-        ru: 'ы',
-        RU: 'Ы',
-        en: 's',
-        EN: 'S',
-        perceivesCapsLk: true,
-    },
-    {
-        keyCode: 'KeyD',
-        ru: 'в',
-        RU: 'В',
-        en: 'd',
-        EN: 'D',
-        perceivesCapsLk: true,
-    },
-    {
-        keyCode: 'KeyF',
-        ru: 'а',
-        RU: 'А',
-        en: 'f',
-        EN: 'F',
-        perceivesCapsLk: true,
-    },
-    {
-        keyCode: 'KeyG',
-        ru: 'п',
-        RU: 'П',
-        en: 'g',
-        EN: 'G',
-        perceivesCapsLk: true,
-    },
-    {
-        keyCode: 'KeyH',
-        ru: 'р',
-        RU: 'Р',
-        en: 'h',
-        EN: 'H',
-        perceivesCapsLk: true,
-    },
-    {
-        keyCode: 'KeyJ',
-        ru: 'о',
-        RU: 'О',
-        en: 'j',
-        EN: 'J',
-        perceivesCapsLk: true,
-    },
-    {
-        keyCode: 'KeyK',
-        ru: 'л',
-        RU: 'Л',
-        en: 'k',
-        EN: 'K',
-        perceivesCapsLk: true,
-    },
-    {
-        keyCode: 'KeyL',
-        ru: 'д',
-        RU: 'Д',
-        en: 'l',
-        EN: 'L',
-        perceivesCapsLk: true,
-    },
-    {
-        keyCode: 'Semicolon',
-        ru: 'ж',
-        RU: 'Ж',
-        en: ';',
-        EN: ':',
-        perceivesCapsLk: true,
-    },
-    {
-        keyCode: 'Quote',
-        ru: 'э',
-        RU: 'Э',
-        en: "'",
-        EN: '"',
-        perceivesCapsLk: true,
-    },
-    {
-        keyCode: 'Enter',
-        ru: 'Enter',
-        RU: 'Enter',
-        en: 'Enter',
-        EN: 'Enter',
-        perceivesCapsLk: false,
-    },
-    {
-        keyCode: 'ShiftLeft',
-        ru: 'Shift',
-        RU: 'Shift',
-        en: 'Shift',
-        EN: 'Shift',
-        perceivesCapsLk: false,
-    },
-    {
-        keyCode: 'KeyZ',
-        ru: 'я',
-        RU: 'Я',
-        en: 'z',
-        EN: 'Z',
-        perceivesCapsLk: true,
-    },
-    {
-        keyCode: 'KeyX',
-        ru: 'ч',
-        RU: 'Ч',
-        en: 'x',
-        EN: 'X',
-        perceivesCapsLk: true,
-    },
-    {
-        keyCode: 'KeyC',
-        ru: 'с',
-        RU: 'С',
-        en: 'c',
-        EN: 'C',
-        perceivesCapsLk: true,
-    },
-    {
-        keyCode: 'KeyV',
-        ru: 'м',
-        RU: 'М',
-        en: 'v',
-        EN: 'V',
-        perceivesCapsLk: true,
-    },
-    {
-        keyCode: 'KeyB',
-        ru: 'и',
-        RU: 'И',
-        en: 'b',
-        EN: 'B',
-        perceivesCapsLk: true,
-    },
-    {
-        keyCode: 'KeyN',
-        ru: 'т',
-        RU: 'Т',
-        en: 'n',
-        EN: 'N',
-        perceivesCapsLk: true,
-    },
-    {
-        keyCode: 'KeyM',
-        ru: 'ь',
-        RU: 'Ь',
-        en: 'm',
-        EN: 'M',
-        perceivesCapsLk: true,
-    },
-    {
-        keyCode: 'Comma',
-        ru: 'б',
-        RU: 'Б',
-        en: ',',
-        EN: '<',
-        perceivesCapsLk: true,
-    },
-    {
-        keyCode: 'Period',
-        ru: 'ю',
-        RU: 'Ю',
-        en: '.',
-        EN: '>',
-        perceivesCapsLk: true,
-    },
-    {
-        keyCode: 'Slash',
-        ru: '.',
-        RU: ',',
-        en: '/',
-        EN: '?',
-        perceivesCapsLk: false,
-    },
-    {
-        keyCode: 'ArrowUp',
-        ru: '↑',
-        RU: '↑',
-        en: '↑',
-        EN: '↑',
-        perceivesCapsLk: false,
-    },
-    {
-        keyCode: 'ShiftRight',
-        ru: 'Shift',
-        RU: 'Shift',
-        en: 'Shift',
-        EN: 'Shift',
-        perceivesCapsLk: false,
-    },
-    {
-        keyCode: 'ControlLeft',
-        ru: 'Ctrl',
-        RU: 'Ctrl',
-        en: 'Ctrl',
-        EN: 'Ctrl',
-        perceivesCapsLk: false,
-    },
-    {
-        keyCode: 'MetaLeft',
-        ru: 'Win',
-        RU: 'Win',
-        en: 'Win',
-        EN: 'Win',
-        perceivesCapsLk: false,
-    },
-    {
-        keyCode: 'AltLeft',
-        ru: 'Alt',
-        RU: 'Alt',
-        en: 'Alt',
-        EN: 'Alt',
-        perceivesCapsLk: false,
-    },
-    {
-        keyCode: 'Space',
-        ru: ' ',
-        RU: ' ',
-        en: ' ',
-        EN: ' ',
-        perceivesCapsLk: false,
-    },
-    {
-        keyCode: 'AltRight',
-        ru: 'Alt',
-        RU: 'Alt',
-        en: 'Alt',
-        EN: 'Alt',
-        perceivesCapsLk: false,
-    },
-    {
-        keyCode: 'ArrowLeft',
-        ru: '←',
-        RU: '←',
-        en: '←',
-        EN: '←',
-        perceivesCapsLk: false,
-    },
-    {
-        keyCode: 'ArrowDown',
-        ru: '↓',
-        RU: '↓',
-        en: '↓',
-        EN: '↓',
-        perceivesCapsLk: false,
-    },
-    {
-        keyCode: 'ArrowRight',
-        ru: '→',
-        RU: '→',
-        en: '→',
-        EN: '→',
-        perceivesCapsLk: false,
-    },
-    {
-        keyCode: 'ControlRight',
-        ru: 'Ctrl',
-        RU: 'Ctrl',
-        en: 'Ctrl',
-        EN: 'Ctrl',
-        perceivesCapsLk: false,
-    },
-];
-
-export default keyBtn;
